requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.17.0
fredapi>=0.5.1

